+++ compiler-plugin/src/main/kotlin/io/tlaloc/plugin/TlalocCompilerPluginRegistrar.kt
@@ -78,6 +78,7 @@ class TlalocCompilerPluginRegistrar : CompilerPluginRegistrar() {
                 dumpGradSourceDir = dumpDir,
                 options = options,
                 handoff = handoff,
+                configuration = configuration,
             ),
         )
     }